Implementing metrics policy resource
